+++ classes/WP_CLI/Action/Cancel_Command.php
