+++ classes/WP_CLI/Action/Delete_Command.php
