+++ classes/WP_CLI/Action/Generate_Command.php
