+++ classes/WP_CLI/Action/Next_Command.php
