+++ classes/WP_CLI/Action_Command.php
@@ -32,7 +32,7 @@ class Action_Command extends \WP_CLI_Command {
 	 * @return void
 	 */
 	public function cancel( array $args, array $assoc_args ) {
-		require_once 'Action/Cancel.php';
+		require_once 'Action/Cancel_Command.php';
 		$command = new Action\Cancel_Command( $args, $assoc_args );
 		$command->execute();
 	}
@@ -84,7 +84,7 @@ class Action_Command extends \WP_CLI_Command {
 	 * @return void
 	 */
 	public function create( array $args, array $assoc_args ) {
-		require_once 'Action/Create.php';
+		require_once 'Action/Create_Command.php';
 		$command = new Action\Create_Command( $args, $assoc_args );
 		$command->execute();
 	}
@@ -116,7 +116,7 @@ class Action_Command extends \WP_CLI_Command {
 	 * @return void
 	 */
 	public function delete( array $args, array $assoc_args ) {
-		require_once 'Action/Delete.php';
+		require_once 'Action/Delete_Command.php';
 		$command = new Action\Delete_Command( $args, $assoc_args );
 		$command->execute();
 	}
@@ -165,7 +165,7 @@ class Action_Command extends \WP_CLI_Command {
 	 * @return void
 	 */
 	public function generate( array $args, array $assoc_args ) {
-		require_once 'Action/Generate.php';
+		require_once 'Action/Generate_Command.php';
 		$command = new Action\Generate_Command( $args, $assoc_args );
 		$command->execute();
 	}
@@ -203,7 +203,7 @@ class Action_Command extends \WP_CLI_Command {
 	 * @return void
 	 */
 	public function get( array $args, array $assoc_args ) {
-		require_once 'Action/Get.php';
+		require_once 'Action/Get_Command.php';
 		$command = new Action\Get_Command( $args, $assoc_args );
 		$command->execute();
 	}
@@ -261,7 +261,7 @@ class Action_Command extends \WP_CLI_Command {
 	 * @subcommand list
 	 */
 	public function subcommand_list( array $args, array $assoc_args ) {
-		require_once 'Action/List.php';
+		require_once 'Action/List_Command.php';
 		$command = new Action\List_Command( $args, $assoc_args );
 		$command->execute();
 	}
@@ -313,7 +313,7 @@ class Action_Command extends \WP_CLI_Command {
 	 * @return void
 	 */
 	public function next( array $args, array $assoc_args ) {
-		require_once 'Action/Next.php';
+		require_once 'Action/Next_Command.php';
 		$command = new Action\Next_Command( $args, $assoc_args );
 		$command->execute();
 	}
@@ -345,7 +345,7 @@ class Action_Command extends \WP_CLI_Command {
 	 * @return void
 	 */
 	public function run( array $args, array $assoc_args ) {
-		require_once 'Action/Run.php';
+		require_once 'Action/Run_Command.php';
 		$command = new Action\Run_Command( $args, $assoc_args );
 		$command->execute();
 	}
